Boost: Refactor how Data Sync Entries are initialized (#41673)
