+++ projects/packages/wp-js-data-sync/changelog/update-boost-cleanup-wp-js-data-sync
@@ -1,0 +1,4 @@
+Significance: patch
+Type: added
+
+Add new helper function for registering read only entries
